Move schema manager console to separate project
Makes for easier downstream overloading of services.
