+++ src/Microsoft.Health.Dicom.SchemaManager.Console/Properties/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+using System.Resources;
+
+[assembly: NeutralResourcesLanguage("en-us")]
+[assembly: CLSCompliant(false)]
