+++ src/Microsoft.Health.Dicom.SchemaManager.Console/Properties/launchSettings.json
@@ -1,0 +1,8 @@
+{
+  "profiles": {
+    "Microsoft.Health.Dicom.SchemaManager": {
+      "commandName": "Project",
+      "commandLineArgs": "apply --connection-string \"server=(local);Initial Catalog=DICOM3;TrustServerCertificate=True;Integrated Security=True\" --version 20"
+    }
+  }
+}
