+++ src/Microsoft.Health.Dicom.SchemaManager/SchemaManagerParserBuilder.cs
@@ -1,0 +1,26 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+using System.CommandLine;
+using System.CommandLine.Builder;
+using System.CommandLine.Parsing;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Microsoft.Health.Dicom.SchemaManager;
+
+public static class SchemaManagerParser
+{
+    public static Parser Build(ServiceProvider serviceProvider)
+    {
+        var commandLineBuilder = new CommandLineBuilder();
+
+        foreach (Command command in serviceProvider.GetServices<Command>())
+        {
+            commandLineBuilder.Command.AddCommand(command);
+        }
+
+        return commandLineBuilder.UseDefaults().Build();
+    }
+}
